+++ spec/twinfield/payment_transaction_spec.rb
@@ -1,25 +1,25 @@
 require 'spec_helper'
 
-describe Twinfield::PaymentTransaction do
+describe Twinfield::Transaction do
   include SessionStubs
   include ProcessxmlStubs
 
   let(:payment_transaction) do
-    trans = Twinfield::PaymentTransaction.new(code: "PIN", currency: "EUR", date: Date.new(2021,1,1))
-    trans.lines << Twinfield::PaymentTransaction::Line.new(type: :total, balance_code: "1230", value: 0.0, debitcredit: :debit)
-    trans.lines << Twinfield::PaymentTransaction::Line.new(type: :detail, balance_code: 1300, value: 60.5, debitcredit: :credit, customer_code: 1003, invoicenumber: 14)
-    trans.lines << Twinfield::PaymentTransaction::Line.new(type: :detail, balance_code: 1234, value: 60.5, debitcredit: :debit)
+    trans = Twinfield::Transaction.new(code: "PIN", currency: "EUR", date: Date.new(2021,1,1))
+    trans.lines << Twinfield::Transaction::Line.new(type: :total, balance_code: "1230", value: 0.0, debitcredit: :debit)
+    trans.lines << Twinfield::Transaction::Line.new(type: :detail, balance_code: 1300, value: 60.5, debitcredit: :credit, customer_code: 1003, invoicenumber: 14)
+    trans.lines << Twinfield::Transaction::Line.new(type: :detail, balance_code: 1234, value: 60.5, debitcredit: :debit)
     trans
   end
 
-  describe Twinfield::PaymentTransaction::Line do
+  describe Twinfield::Transaction::Line do
     describe "#to_xml" do
       it "converts a detail line to xml" do
-        expect(Twinfield::PaymentTransaction::Line.new(type: :detail, balance_code: 1055, value: 60.5, debitcredit: :debit).to_xml).to eq("<line type=\"detail\">\n  <dim1>1055</dim1>\n  <value>60.5</value>\n  <description/>\n  <debitcredit>debit</debitcredit>\n</line>")
-        expect(Twinfield::PaymentTransaction::Line.new(type: :detail, balance_code: 8020, value: 50, debitcredit: :credit, vatcode: "VH").to_xml).to eq("<line type=\"detail\">\n  <dim1>8020</dim1>\n  <value>50.0</value>\n  <description/>\n  <debitcredit>credit</debitcredit>\n  <vatcode>VH</vatcode>\n</line>")
+        expect(Twinfield::Transaction::Line.new(type: :detail, balance_code: 1055, value: 60.5, debitcredit: :debit).to_xml).to eq("<line type=\"detail\">\n  <dim1>1055</dim1>\n  <value>60.5</value>\n  <description/>\n  <debitcredit>debit</debitcredit>\n</line>")
+        expect(Twinfield::Transaction::Line.new(type: :detail, balance_code: 8020, value: 50, debitcredit: :credit, vatcode: "VH").to_xml).to eq("<line type=\"detail\">\n  <dim1>8020</dim1>\n  <value>50.0</value>\n  <description/>\n  <debitcredit>credit</debitcredit>\n  <vatcode>VH</vatcode>\n</line>")
       end
       it "converts a total line to xml" do
-        expect(Twinfield::PaymentTransaction::Line.new(type: :total, balance_code: "0000", value: 0.0, debitcredit: :debit).to_xml).to eq("<line type=\"total\">\n  <dim1>0000</dim1>\n  <value>0.0</value>\n  <debitcredit>debit</debitcredit>\n</line>")
+        expect(Twinfield::Transaction::Line.new(type: :total, balance_code: "0000", value: 0.0, debitcredit: :debit).to_xml).to eq("<line type=\"total\">\n  <dim1>0000</dim1>\n  <value>0.0</value>\n  <debitcredit>debit</debitcredit>\n</line>")
       end
     end
   end
@@ -67,7 +67,7 @@ describe Twinfield::PaymentTransaction do
 
   describe "#to_transaction_match_line_xml" do
     it "returns a valid piece of xml" do
-      xml = Nokogiri::XML(Twinfield::PaymentTransaction.new(code: "PIN", number: "20210121").to_transaction_match_line_xml(2))
+      xml = Nokogiri::XML(Twinfield::Transaction.new(code: "PIN", number: "20210121").to_transaction_match_line_xml(2))
       xml.css("line transcode").text == "VRK"
       xml.css("line transnumber").text == "20210120"
       xml.css("line transline").text == "1"