+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All changes to the project will be documented in this file.
 ## Pending release
 
 * [BREAKING] Invoices are now what Twinfield calls SalesInvoices, not SalesTransactions
-* [FEATURE] Added different classes that act more like ruby native objects in the Twinfield domain, such as Twinfield::Customer, Twinfield::Invoice and Twinfield::Browse::Transaction::Customer
+* [FEATURE] Added different classes that act more like ruby native objects in the Twinfield domain, such as Twinfield::Customer, Twinfield::Invoice, Twinfield::Transaction and Twinfield::Browse::Transaction::Customer
 * [FEATURE] Allowed for OAuthSession; which is the preferred way of creating sessions
 * [BUGFIX] Finder code made to work